Added HAL and LL solution

+++ F103_Buttons_LL/Core/Inc/btnFcn.h
@@ -1,5 +1,5 @@
 /** 
- * Project: F103_Buttons_HAL
+ * Project: F103_Buttons_LL
  * 
  * @file btnFcn.h
  * 
@@ -40,7 +40,7 @@
  *
  * @return 0 while button pressed, 1 while button not pressed
  */
-uint8_t BtnRelease(uint8_t *btn, GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin, void (*BtnFcn)(void));
+uint8_t BtnRelease(uint8_t *btn, GPIO_TypeDef *GPIOx, uint32_t GPIO_Pin, void (*BtnFcn)(void));
 
 
 /**@brief Routine for pressing button switch
@@ -52,6 +52,6 @@ uint8_t BtnRelease(uint8_t *btn, GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin, void (*
  *
  * @return 0 while button pressed, 1 while button not pressed
  */
-uint8_t BtnPress(uint8_t *btn, GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin, void (*BtnFcn)(void));
+uint8_t BtnPress(uint8_t *btn, GPIO_TypeDef *GPIOx, uint32_t GPIO_Pin, void (*BtnFcn)(void));
 
 /******************************** END OF FILE *******************************/
